Inject refund subcommand extension
